// interfaces/order.interface.ts
export interface IOrder {
    amount:                   number;
    bank:                     string;
    bank_account:             string;
    customer_document_number: string;
    customer_email:           string;
    customer_full_name:       string;
    customer_phone_number:    string;
    inverted:                 string;
    referrals_reference:      string;
}